src/modules/accounts/useCases/sendForgotPassWordMail/SendForgotPassWordMailUseCase.ts
import { IUserRepository } from "@modules/accounts/repositories/IUserRepository";
import { IUsersTokensRepository } from "@modules/accounts/repositories/IUsersTokens";
import { IDateProvider } from "@shared/container/providers/DateProvider/IDateProvider";
import { IMailProvider } from "@shared/container/providers/MailProvider/IMailProvider";
import { AppError } from "@shared/errors/AppError";
import { inject, injectable } from "tsyringe";
import { v4 as uuidV4 } from "uuid";
import { resolve } from "path";

@injectable()
class SendForgotPassWordMailUseCase {
    constructor(
        @inject("UserRepository")
        private userRepository: IUserRepository,
        @inject("UsersTokensRepository")
        private usersTokensRporitory: IUsersTokensRepository,
        @inject("DayjsDateProvider")
        private dateProvider: IDateProvider,
        @inject("MailProvider")
        private mailProvider: IMailProvider
    ) {}
    async execute(email: string) {
        const user = await this.userRepository.findByEmail(email);

        if (!user) {
            throw new AppError("User not exist");
        }

        const expires_date = await this.dateProvider.addHours(3);
        const token = uuidV4();

        await this.usersTokensRporitory.create({
            user_id: user.id,
            refresh_token: token,
            expires_date,
        });

        const path = resolve(
            __dirname,
            "..",
            "..",
            "views",
            "emails",
            "forgotPassword.hbs"
        );

        const variables = {
            name: user.name,
            link: `${process.env.FORGOT_MAIL_URL}${token}`,
        };

        await this.mailProvider.sendMail(
            email,
            "recuperação de senha",
            variables,
            path
        );
    }
}

export { SendForgotPassWordMailUseCase };
